+++ web/components/providers/imc-provider.tsx
@@ -814,6 +814,24 @@ export default function InterModuleCommunicationProvider({
           return editorContext?.persistSettings?.locale ?? "en";
         },
       ],
+      [
+        IMCMessageTypeEnum.EditorArtifactUpdate,
+        async (
+          senderWindow: Window,
+          message: IMCMessage,
+          abortSignal?: AbortSignal,
+        ) => {
+          const artifact = message.payload;
+
+          editorContext?.updateModalStates({
+            artifact: {
+              isOpen: true,
+              artifact: artifact,
+              fromViewId: message.from,
+            },
+          });
+        },
+      ],
     ]);
 
     return newMap;