+++ npm-packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pulse-editor/shared-utils
 
+## 0.1.1-beta.72
+
+### Patch Changes
+
+- Add artifact output hook
+
 ## 0.1.1-beta.71
 
 ### Patch Changes